Move tracking group resolvers into dedicated modules
