+++ back/boxtribute_server/mobile_distribution/tracking_group/fields.py
@@ -1,0 +1,41 @@
+from ariadne import ObjectType
+
+from ...authz import authorize
+from ...models.definitions.distribution_event import DistributionEvent
+from ...models.definitions.distribution_event_tracking_log_entry import (
+    DistributionEventTrackingLogEntry,
+)
+
+distribution_events_tracking_group = ObjectType("DistributionEventsTrackingGroup")
+
+
+@distribution_events_tracking_group.field("distributionEventsTrackingEntries")
+def resolve_distribution_events_tracking_group_tracking_entries(
+    distribution_events_tracking_group_obj, _
+):
+    authorize(
+        permission="distro_event:read",
+        base_id=distribution_events_tracking_group_obj.base_id,
+    )
+    return DistributionEventTrackingLogEntry.select().where(
+        (
+            DistributionEventTrackingLogEntry.distro_event_tracking_group_id
+            == distribution_events_tracking_group_obj.id
+        )
+    )
+
+
+@distribution_events_tracking_group.field("distributionEvents")
+def resolve_distribution_events_tracking_group_distribution_events(
+    distribution_events_tracking_group_obj, _
+):
+    authorize(
+        permission="distro_event:read",
+        base_id=distribution_events_tracking_group_obj.base_id,
+    )
+    return DistributionEvent.select().where(
+        (
+            DistributionEvent.distro_event_tracking_group_id
+            == distribution_events_tracking_group_obj.id
+        )
+    )
